Move `LoginRoute` into own module, base on `session`.
